Add helper function to insert default file information (size, filename and the file)
